requirements.txt
streamlit==1.33.0
openai==1.99.9
python-dotenv==1.0.1
requests==2.32.4
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.17

